refactor: speech_bubble.py 纯函数区抽出（批6-2）
- 分页/定位等 9 个纯函数 + 常量 → speech_bubble_text.py
- PetSpeechBubble 绘制代码一行未动（防视觉回归）
- re-export 保持全部调用点兼容

结构优化批6-2（STRUCT_PLAN_FINAL.md）
